Edited ddl bigint to int, installed mariadb and dotenv for backend, created pool connection to database and function with test query

+++ backend/db/0_DDL.sql
@@ -3,10 +3,10 @@
 CREATE TABLE `journeys` (
   `departure_date_time` datetime DEFAULT NULL,
   `return_date_time` datetime DEFAULT NULL,
-  `departure_station_id` bigint(20) DEFAULT NULL,
+  `departure_station_id` int DEFAULT NULL,
   `departure_station_name` varchar(100) DEFAULT NULL,
-  `return_station_id` bigint(20) DEFAULT NULL,
+  `return_station_id` int DEFAULT NULL,
   `return_station_name` varchar(100) DEFAULT NULL,
-  `covered_distance_m` bigint(20) DEFAULT NULL,
-  `duration_sec` bigint(20) DEFAULT NULL
+  `covered_distance_m` int DEFAULT NULL,
+  `duration_sec` int DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
